Add first version of a register table abstraction
The idea here is to allow the definition of an external interface for an
embedded system as a chunk of linear memory. The application side should be
able to trivially interact with the contents of that table. On the other side
of the picture the communication system of the system should have functions to
easily read and write blocks of register data.

Another driving idea behind this implementation is to be able to express an
entire, even complex, register table in one C language expression in order to
keep things tidy.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>