Implement block_write() API
This is a little more complex than the block_read() API.

It also works across hole-free area borders. But in addition it checks if the
block write will result in a viable register table. It does this by iterating
over touched entries and produce what they would look like. After that it will
call the deserialisers and validators on the produced items and report any
errors.

Only block writes that pass all checks will be carried out.

This introduces a couple of utilities, one of which block_read() now uses too.
In fact, that API (register_block_touches_hole) is now part of the public API
because this also adds _unsafe() variants of read and write for cases in which
the application really needs to to that. And register_block_touches_hole()
makes it possible to actually use these.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>