Get rid of string type registers
Strings are a little tougher than the other types in this implementation,
because it is not a scalar type and therefore its size is not known. It is
certainly possible to support strings, but I am not at all convinced by their
usefulness. If need be, fixed length strings can be represented by sequences of
unsigned integer registers.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>