Fix initialisation code when setting defaults
So... when requiring that register_set() only works with initialised register
tables, it's tough to use it in the initialisation procedure itself. This now
sets the INITIALISED bit a tad earlier and clears it in case of an error.

What register_set() really needs is the t->entries count to be valid. And that
is the case in the current code.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>