registers: Wrap public API in extern "C" { ... }
The module isn't really supposed to be compiled with a C++ compiler. The module
itself (register-table.c) targets C99. When used in C++ applicaitons, register
tables should be defined in a small C99 unit, since some of the macros defined
in the module depend on C99 features.

This change however, makes it possible to *use* most of the module's public API
in C++ code. That is fine in terms of interoperability.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>