Add utility to set memory based on hex-digit-string
The rationale here is, that a common use-case for a string of memory is the
reflection of the state of a git-controlled source tree. One of the pivotal
data-points with this problem is the checksum of the currently checked out
commit. This may lead to a pre-processor definition such as:

    #define GIT_HASH "123456789abcdef"

The new API can convert this ASCII string into a string of uint16_t values,
that live inside a register table:

    register_set_from_hexstr(&regs, addr, sha1, 12u);

Which would lead to the memory starting at ‘addr’ to look like this:

    [ 0x1234 0x5678 0x9abc ]

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>