Add binary-format module as (de)serialisation primitives
These are useful when reading from/writing to memory with the intend to share
the result with another machine. In other cases, internal data represenations
should never matter. In order to generate data for external protocols, these
functions are supposed to provide unsigned integers, twos-complement signed
integers and ieee754 floating point numbers in single (32 bit) and double (64
bit) precision.

The API is fairly simple and inlinable, so it shouldn't have an impact compared
to hand-written code. The internals make use of the SYSTEM_ENDIANNESS_* macros,
in order to determine the machine's internal endianness. If these are not set,
it falls back to the usual mask-shift-or routine.

The code currently assumes of the underlying machine:

a. Twos compliment signedness.
b. IEEE-754 floating point numbers.
c. Octet-order works in a straight-forward manner with floating point values.
d. The existence of 32 and 64 bit wide floating point numbers.

Point a is a fair assumption. Point b is more problematic, but I don't know a
stable way to detect the availability. Point c is also less straight-forward
than one might imagine and again, I don't know how to detect this, especially
in a cross-compilation environment. Point d is also a problem: There are some
plattforms that don't have double precision floats and some alias double to
float — again, that's tough to tell at compile-time with cross-compilation in
place. I should probably find a way to handle point d, even if only at run
time.

Another way would be to add a sort of plattform capability data base to the
SetupTargetCPU CMake module. We're sort of doing that already with the endi-
anness detection via SYSTEM_ENDIANNESS_*.

Yet another way would be to build and run the test-suite on real hardware. The
test-anything-protocol lends itself to that sort of thing; just output the test
protocol to a serial port, read it on a host and parse using TAP::Parse.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>