Add register-utilities API
The idea here is to get pretty-printers for a couple of types used in the
library. The reason for having them is this: The register-table code runs a
whole slew of tests when initialising an instance. Those are important to
guarantee the integrity of the table and its data. In case the initialisation
step fails, there is little to nothing a self-contained piece of software is
able to do to mitigate the situation.

Unfortunately, there is no way of detecting problems like this at compile time.
It is possible, however, to use the same register-table definition that is used
in the actual software, in a unit test and just run the initialisation step at
test-time.

To achieve this, it is convenient to keep a register-table definition in a
separate source file that can be compiled as part of the actual software and
also included in a test-suite file. With that, this becomes possible:

  #include <stdlib.h>
  #include <c/register-table.h>
  #include <c/register-utilities.h>
  #include <common/compiler.h>
  #include <tap.h>
  #include "version.h"
  #include "../src/task/registers.h"
  #include "../src/task/registers.c"
  int
  main(UNUSED int argc, UNUSED char *argv[])
  {
      plan(1);
      RegisterInit bootsuccess = register_init(&registers);
      if (bootsuccess.code != REG_INIT_SUCCESS) {
          register_table_print(stdout, "# ", &registers);
          printf("#\n");
          register_init_print(stdout, "# ", bootsuccess);
      }
      ok(bootsuccess.code == REG_INIT_SUCCESS, "RegisterTable initialises");
      return EXIT_SUCCESS;
  }

Feedback is pretty quick, if your test-suite runs as part of your normal build
pipeline then.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>