Add persistent storage API
A common task in embedded systems is to store persistent data in non-volatile
memory such as flash memory or EEPROM. This module allows storing such data,
prefixed by a checksum.

The default checksum algorithm is a trivial unsigned 16 bit sum across all
octets of the stored data. The checksum algorithm is configurable to arbitrary
functions that produce 16 or 32 bit checksums.

The storage medium has to allow for block access.

The storage may be placed in arbitrary places around the storage medium.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>