persistency: Remove dependency upon octets
Instead of insisting to use octets with read and write block access to media,
expect the supplied callbacks to work with the system's native idea of bytes.

Most people will not care about this, since on most plattforms a byte is indeed
an octet. On some plattforms that is not the case. Like some DSPs from Texas
Instruments.

This leaves the door open for extending the API to allow overriding the default
checksum width, in terms of addressable word cia the read and write callbacks.
But for now this is good enough and makes the code work naturally on more ar-
chitectures than before.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>