regs: Add API to copy register areas within a register table
Oftentimes, when parts of a register table lives in persistent memory and
others in RAM, it is desirable to transfer persistent parts into RAM upon boot
up. This API is meant to address this usecase.

It requires at least on of the areas to reside in RAM. If both areas were only
accessible via block IO, we'd need an auxiliary buffer or only transfer memory
byte by byte. It's easy to define an API like that, but for now it seems less
useful.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>