Behave when being used in an application for the zephyr rtos
There are different ways of using ufw in zephyr programs: It can be built
outside of the zephyr tree, which means we have to setup compilers on our own;
but it can also be used inside of zephyr and in that case, we should just trust
zephyr's build-system to do the right thing. In the future, maybe ufw should
also be possible to be built as a zephyr module. In that case we also just need
to trust.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>