Remove warning for MISRA 12.13
Here is the rule from the standard:

    Rule 12.13 (advisory):
        The increment (++) and decrement (--) operators should
        not be mixed with other operators in an expression.

The rationale is to not to mix side-effecting statements in the middle of other
statements, to reduce the total complexity of the complete statement.

Unfortunately, a lot of code uses idioms like

    while (*in != '\0')
        *out++ = *in++;

which causes warnings due to this rule. Since this is an advisory rule, it gets
disabled for now.

Signed-off-by: Martin Gritzan <martin.gritzan@gmail.com>