Remove warning for MISRA 15.3
Similarly to the 12.13 change previously, there is lots of existing code,
violating this one. The rule is required by MISRA, but this disables it
anyway, because this is used on code that cannot be fixed as a whole right
now. The rule from the standard:

    Rule 15.3 (required):
        The final clause of a switch statement shall be
        the default clause.

    Rationale:
        The requirement for a final default clause is defensive programming.
        This clause shall either take appropriate action or contain a suitable
        comment as to why no action is taken.

This is all part of section 15 of MISRA, which introduces this:

    Rule 15.0 (required):
        The MISRA C switch syntax shall be used.

    Rationale:
        The syntax for the switch statement in C is weak, allowing complex,
        unstructured behaviour. The following text describes the syntax for
        switch statements as defined by MISRA-C and is normative. This, and the
        associated rules, enforce a simple and consistent structure on to the
        switch statement.

This isn't wrong, but existing code bases force this decision.

Signed-off-by: Martin Gritzan <martin.gritzan@gmail.com>