cmake: Be fatally strict for explicit int and fct definition in C
Both of these instances are *NOT* valid since C99. The C99 standard just allows
compilers to tolerate these occurances if they warn about it.

These are probably the single worst features about C, which is why they are now
disallowed if the compiler in question supports it. If not we still try to make
this a warning with *all* C versions.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>