Add foundation for a minimal TAP emitting testing module
The plan is to be able to run test-suites directly on bare-metal systems and
their emulation options without the need for complete operating stacks and HAL
systems involved. This is especially interesting for solutions involving qemu,
that would allow running the current test-suite almost *unchanged*.

On real systems, there are questions that need to be answered, surrounding test
output transmission to the host system that is running the test-harness. I am
leaning toward using semi-hosting here as well (with qemu emulation this should
make things rather straight forward); but I have not worked out all kinks yet.

TAP is a good protocol to use, since it's text-only, human and machine readable
and emitters and harnesses exist for virtually every programming environment
under the sun. Though, bare-metal seems to be a little bit of an outlier.

Libtap (the systems used thus far) doesn't compile trivially on arm-none-eabi
and creating a fork doesn't seem worth it, because its core is rather simple.
Plus I don't like some of its design decisions. Like ‘cmp_ok()’ which takes its
operator as a string argument (which is a cute trick), but uses integers for
the operators argument, which can break when you want to test things that
involve other data types.

I've also looked at cmocka — which at least uses CMake as a build-system — but
it doesn't compile on arm-none-eabi either because of some dependencies. Not
sure if it could be ported, but the whole system looks a little too big for me
to take on that particular task.

So here we go. This implementation only supports deterministic test plans for
now, and its only primitive is ‘ok’. From libtap, I will need at least that and
the functionalities of ‘cmp_ok’ and ‘cmp_mem’ to remove ufw's dependency on it.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>