Suppress warning about non-initialised value
Multiple toolchains report this:

test/memdiff.c: In function ‘memdiff’:
test/memdiff.c:191:83: warning: ‘A.lineoffset’ may be used uninitialized in this function
191 |         printlines(diff, A->lineoffset+diff->columns, (diff->size - A->lineoffset - 1)/diff->columns);
    |                                         ~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~
test/memdiff.c:383:23: note: ‘A.lineoffset’ was declared here
383 |     struct difference A, B;
    |                       ^

This is a false positive, since the code makes sure the ‘lineoffset’ member
is *only* accessed if the ‘valid’ member is set to true. This is a simple fix
to remove the warning.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>