Make it possible to build zephyr applications
While zephyr uses cmake as its build-system, it does do some things differently
than most cmake projects. Also, in contrast to the usual build-process advoca-
ted by ufw, zephyr's build system also only builds exactly one application per
build tree, whereas ufw builds strive to build all application that target a
given board/toolchain/buildcfg in one build-tree.

This attempts to allow for zephyr firmwares to build in such the default ufw
environment. The idea is still for the super-project to define ALL software
that is required for an ecosystem; as opposed to just those dependencies re-
quired for one software/firmware that is part of that ecosystem.

In order for that to work, this implements ‘ufw_add_zephyr’, that lets the user
define all parameters required to build a given external zephyr based firmware.

Example:

    cmake_minimum_required(VERSION 3.13.1)
    project(generic-h7-system NONE)

    list(APPEND CMAKE_MODULE_PATH ${CMAKE_CURRENT_SOURCE_DIR}/ufw/cmake/modules)
    include(SetupUFW)

    ufw_toplevel(
        ROOT      ${CMAKE_SOURCE_DIR}/libraries/ufw
        ARTIFACTS ${CMAKE_BINARY_DIR}/artifacts)

    ufw_add_zephyr(blinky
        KERNEL      ${CMAKE_SOURCE_DIR}/zephyr/kernel
        APPLICATION ${CMAKE_SOURCE_DIR}/applications/sample-basic-blinky
        MODULE_ROOT ${CMAKE_SOURCE_DIR}/zephyr/modules
        MODULES     arm-cmsis hal-stm32
        BUILDTYPES  Release Debug
        BOARDS      stm32h747i_disco_m7
        TOOLCHAINS  gnuarmemb:/usr
        OPTIONS     GENERATE_API_DOCUMENTATION=True)

    ufw_recursive_dispatch()
    enable_language(C)
    enable_language(CXX)
    enable_language(ASM)
    include(InitialiseToolchain)
    initialise_toolchain()
    ufw_subtree_build()

    setup_ufw(libraries/ufw)
    add_subdirectory(applications/sample-basic-blinky)

Note that the CMakeLists.txt file of ‘sample-basic-blinky’ should NOT call
‘find_package(Zephyr …)’ itself, since ‘ufw_recursive_dispatch’ will take care
of it. This is required since third-party libraries, referenced by the super
project, need to be able to determine toolchain details. Notably, ufw's
set_target_cpu function makes this transparent, and translates between the
zephyr world and the usual CMake world that uses "Release", "Debug" as names
for build configurations.

Example CMake file for the application thusly:

    cmake_minimum_required(VERSION 3.13.1)
    project(blinky)
    target_sources(app PRIVATE src/main.c)

A very minimal, yet working, example.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>