Add support for generating system version information
In situations where a given firmware is a submodule of a larger system, it may
be of interest, to know the state of the super-project at the time the system
was built. In order to reflect that this adds a variant of the version header
generation to hold information specific to the super project.

Do this in the toplevel cmake file:

    generate_version_h(
      SYSTEM_VERSION
      TARGET system-version-h
      TEMPLATE "${CMAKE_SOURCE_DIR}/system-version.h.in"
      OUTPUT   "${CMAKE_BINARY_DIR}/system-version.h")

This in the firmware's cmake file:

    if (TARGET system-version-h)
      add_dependencies(my-beautiful-firmware system-version-h)
    endif()

While having this expansion in the firmware's version.h.in:

    #if @@{WITH_SYSTEM_VERSION_H:-0}@@
    #include @@{SYSTEM_VERSION_H:-disabled}@@
    #define WITH_SYSTEM_VERSION_H
    #endif

The latter is now included in the example version.h.in that comes with ufw.
This also adds system-version.h.in that serves as an example template for
super-project version headers.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>