Implement register iteration across a given address range
In more complex systems it may be useful to distribute registers that may have
been changed with a block write operation within the system. This function can
help implementing this.

The implementation uses a callback approach in order to not require heap allo-
cation. Also, determining the first register to call the callback for is cur-
rently implemented by linear search. With short register tables that's probably
good enough. But it would be easy enough to replace this by bisection as well.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>