zephyr: Set ZEPHYR_BASE in addition to Zephyr_ROOT
I still had an old zephyr kernel in ~/src/zephyr and Zephyr's CMake magic was
trying to pick that up. I think they mean well, but it doesn't fit the way that
we're trying to make Zephyr's integration as declarative as possible.

Setting ZEPHYR_BASE in addition to Zephyr_ROOT fixes the issue.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>