Rearrange order of the RegisterType enum
If a register entry is included in the SensorRegister struct but not defined in
the `.entry` of the register table, it is an error case. However, it was not
detected, especially in the bss segment where all the memory is initialized
with 0, because the REG_TYPE_INVALID is 0.

Instead detecting the error and aborting the initialization, the register table
is initialized successfully with the incorrect number of entries. To prevent
this case from NOT being detected as an error, the enum RegisterType is
rearranged in the way that REG_TYPE_INVALID is no longer 0.

Signed-off-by: Jiyeon Hong <jyeons.hong91@gmail.com>