reg,utils: Make printer target opaque and allow exchanging printer
Printing to FILE* is fine for native builds, but in other environments, its not
perfect. For instance, when these pretty printers are to be used within ze-
phyr's shell environment, we'd want to pass struct shell* through. This change
allows this. Since any pointer automatically converts to void*, this should not
break existing code.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>