registers: Fix compilation with newer newlib
With fairly recent newlib (3.3.0), this header has a couple of dependencies
that breaks the availability of macros like PRIu64 (the 64-bit ones). Lea-
ding to compilation errors in ‘register_value_print()’. Including this after
stdio.h fixes the issue.

Seems like a bug in newlib, but this can't be helped. Since the named version
is included in Debian 11, which is an important build plattform.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>