cmake: Remove lots of recursive-cmake cruft
This code was pretty bad. Even for CMake standards.

Previously we've tried to use as little external tooling as possible, and get
by doing everything in a build and code-structure with just git and CMake. Most
things were implementable, thanks to calling CMake recursively and its weird
and wonderful ExternalProject module.

But user friendliness has suffered. A lot. It was very hard to trigger specific
builds. I actually wanted to fix this by somehow making CMake generate a list
of registered build-variants, so I could maybe at least get some sort of shell
completion to pick up on that. But it's all a horrible shitshow.

And the downsides aren't just that. From mmh's ROADMAP:

        This has  several downsides:  The CMakeLists.txt  file becomes
        tough to follow because the  same file gets parsed in multiple
        ways by multiple CMake processes.  The build tree becomes com-
        plex and quite deep so the top-level CMake call can keep track
        of what is happening. This  is rather unergonomic for the user
        to decent into. And finally, it becomes tough to perform para-
        llel build  correctly. If you  use Ninja, for instance  an run
        your top-level ninja in parallel mode, the recursive calls may
        also be  parallel,  which — in complex builds — can  massively
        overload your system.

Combined with the atrocious code quality it's a clear candidate for some-
thing to get rid of. That's where mmh's proposed ‘system’ sub-command comes
into play. The delarative API of our old scheme of specifying the kind of
builds that we'd like to have carried out makes it pretty easy to map this
to a scheme using an external tool that reads such a specification from,
say, a YAML file.

And that is the idea behind that ‘system’ sub-command. The accompanying
branch in the makemehappy space is ‘ft/system-subcmd’.

Note that this is probably not everything that needs to be weeded out. But
you got to start somewhere.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>