Reorganise code-base
Previously the plan was to allow for C and C++ as well as modules common to the
two languages. By now it has become pretty clear, that C++ won't be part of
ufw, though. This gets rid of this separation.

In addition, this adds a ‘ufw/’ prefix to all include files.

Build-system-wise, this removes the ‘ufw++’ and ‘ufw-common’ libraries. For
symmetry with ‘ufw-nosan’, there's now also a ‘ufw-tap-nosan’ variant with
native builds.

This also increases cmake compatibility to 3.13.4.

Clearly, this breaks backward compatibility, so the next release will increment
ufw's major version.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>