Implement per-register user-defined data
To allow users to attach arbitrary data to each register entry of a register
table, this adds a void-pointer to the data type.

To assign to this pointer, there are REGx_...() specifier macros. The old API,
that uses the REG_ prefix, simply assigns NULL to the pointer.

The complete set of REG*_...() macros can now be generated be executing a
program: tools/make-register-macros — this is done to limit errors introduced
by sloppy editing.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>