Add register_user_init API
This allows for users to run their own initialisation code given the existence
of the per-register user data.

We can't easily put this into register_init(), because there is no way for us
to determine whether a function pointer provided by the RegisterTable structure
is valid or not. It could have been done by changing register_init()'s API,
adding the function pointer, or by adding a new function, that does what that
function does plus run this iteration.

Instead this adds a new function that can be called immediately after
register_init(), which achieves the same result.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>