Implement a additional default sources and sinks
Previously sources and sink that are driven by octet buffers were implemented.
This adds /dev/null and /dev/zero like instanes as well file-descriptor driven
instances backed by POSIX read() and write().

In addition, this adds sink and source instances, similar to the buffer driven
one, that is also instrumentable in a way to fail predictably, which is useful
for implementing tests.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>