Add variable length integer encoding/decoding
This is the same variable length encoding as used in google's protobuf. The
most significant bit of each octet of a number denotes whether or not the next
octet is part of the number as well.

This implements unsigned 64 and 32 bit as well as two's complement encoders and
decoders of the same width. The API offers octet-buffer and source/sink based
functions. For the time being, this does NOT implement ZigZag encoded signed
integers.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>