endpoints: Modify API with respect to error codes
Previously, this system used zero as a special case to communicate an endpoint
not being able to produce or consume requested data. The underlying drivers
would return -EOVERFLOW to cause this, which in turn required users of the
endpoint to switch back to an error code to inform its users about problems.

This switch to -ENOMEM for sinks that can't accept the desired amount of data
and -ENODATA for sources that can't produce the desired amount of data.

This also replaces -EILSEQ instead of -EBADMSG for illegal byte sequences in
the SLIP and variable-length-integer implementations.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>