tap: Add okx(), a description-less variant of ok()
There are occasions, where the expression passed into ok() is expressive enough
to fully describe the test that is performed. With the implementation of ok(),
using __VA_ARGS__, you are not allowed to pass zero arguments into this list.

Adding a variant, that does not use __VA_ARGS__ but instead passes NULL into
the tap_test_ok()'s format parameter, makes it easy to print the expression
instead of a printf()-like format.

It is not safe to pass the expression into tap_result as a format string,
because it might contain percent-sequences if properly crafted.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>