rfc1055: Fix decoder when EOF is lost in inopportune places
When a EOF is lost in the decoding scheme, that uses EOF as a start of frame
delimiter as well, previously we'd lock onto the next EOF as the start of frame
delimiter. But since we lost one before the next octet will be another EOF,
which will look like an empty frame. Then the actual frame will be treated as
noise, since we are looking for SOF again, so the next frame will look empty
also. …and so on.

The fix is to assume that when we're looking for start-of-frame, only a EOF
octet is acceptable. If that's not the case, we switch to scanning for end of
frame. This will resynchronise eventually, because an intact frame will end in
EOF. And then the next intact frame after that will start in EOF, which means
we're synchronised again.

There really is not much upside to using start-of-frame, if the same octet as
the end-of-frame delimiter is used…

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>