bitops: Rename BITS_PER_* macros to UFW_BITS_PER_*
There are vendors like Zephyr, that also USE BITS_PER_* macros. And because
there are no namespaces in C, you get conflicts when using headers that in any
combination use both.

Previously, we added some detection that allowed us to still use the old names,
when the right order of inclusions is used. This works but is clumsy. So now,
we will use a UFW_ prefix for these symbols in order to limit symbol conflicts.

This is technically an API break, and thus the next release will be 4.0.0.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>