binary: Revise binary-format module
The main issue with the old implementation is that it used the following
pattern to make native octet-order data accesses:

    *(uint64_t *)src;  /* src is a pointer to arbitrary memory */

This works on some architectures, but is not generally portable. Data types
have alignment rules, and if an access like that attempted at a misaligned
address, bus-errors are possible, crashing your application.

The new implementation uses unrolled loops that copy data byte by byte.

This is the naïve way, and it would be possible to detect if the short cut
works, but in general, we'd need to perform a modulo operation. If it was known
that alignment rules always yield a power-of-two, the same could be achieved
with a simple bit-operation. In general, that's not possible, since the C
standard does not mention powers of two being returned by "alignof(…)".

With modulo, it's probably not worth doing this, because on many architectures,
this is a very expensive operation. We may make this optional in the future.
For stdalign.h, ufw would have to upgrade its C standard to C11 at least as
well, reducing its applicability.

The code is generated now, since all of the binary format code is very
mechanical in nature. This should lead to very simple, single branch code for
the compiler to chew on.

With this in place, we may add non power-of-two accessors to the module as
well (for instance bf_ref_u40l(…) and the like). This is a little more
difficult to get right on systems where the a byte is not an octet.

For the minute, this update mostly reproduces the old functionality.

The rewrite does however break API compatibility in some places: More arguments
were made ‘const’ (this is probably invisible to the user) and the setter
functions now return the memory address after the last piece of memory that was
touched by the storing operation.

Also, the previous use of integer pointers being used to point to memory was
changed to now employ void pointers, as is more common in C.

If the return value of the setter function was used in existing code, this is
an incompatibility that needs to be addresses. If that is not the case, these
changes will be transparent to the user.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>