build: Get rid of CheckTypeSize.cmake
As it turns out, this CMake provided module does this: Compile some code, feed
the binary into the "strings" program and grep for some known patterns. This
doesn't work very well with toolchains outside of GCC and Clang. Oh well.

What we really want to know here is whether or not a data type is available in
a toolchain, because of the absence of "uint8_t" on C2000 DSPs.

For C2000 this worked out by accident, since the broken test appears to make
things work because ufw will assume uint8_t is not available in the toolchain.
And for C2000, that's correct. But CMake only makes it appear that way, because
its detection for this toolchains is inadequate.

Since ufw is also used with the ti-arm toolchains, this breaks down, because
the detection doesn't work on those either, which means that on this platform,
ufw thought uint8_t wasn't available either.

Since we're not really interested in the size of the type here, I am replacing
that module with a quick check_c_source_compiles() wrapped in a macro.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>