cmake: Add ufw_install utility
This is a wrapper around ‘install(…)’, which currently only sets the CMake
install component to the value of ${UFW_INSTALL_COMPONENT}. This is useful
because in the ufw/mmh environment we are using CMake's installation sub-
system to populate an artefacts tree. But we don't want to invoke random
third-party installation routines, that may even be broken (picolibc, I am
looking at you!).

ufw_install() is a drop-in replacement for install(). gitint_install() already
uses this component setup.