cmake: Invalidise variable before use
I don't know what CMake is doing here. When a previous call to
ufw_test_a_harness() found a program before (so found was set) but the program
turned out to be unusable, so that the function would be called with a
different program (first tap-harness, then prove), then for some reason, the
‘found’ variable was still set, even though this is a different function call.
Then it would just perform the previous tests again, claiming that the second
tool would be unusable as well.

Doing this solves this, even though it feels really wrong. Just CMake things.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>