cmake, test: Support multiple harness feature-sets
Namely perl-prove and guile-tap-harness. The former is the classic TAP harness
from Perl and its command line API. The latter is another TAP harness that
started out as a possible drop-in replacement for prove, but since deviated,
planning to add interesting features for use in embedded development, like
reading SLIPed TAP data from stdin, which is interesting for running embedded
test-suites on real hardware.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>