regp: Add a specification for the proposed register protocol
After some research, the story in terms for simple protocols to interface
embedded systems by seems rather sorry. The best candidate thus far seemed
CoAP, which is not terribly simple, however. However the current state of its
implementation in Zephyr is pretty tightly tied to its networking subsystem,
which makes it difficult to use in non-networking channels, with standard CoAP
being specified for UDP channels. Also implementing CoAP is not all too
straight forward, depending on unicode encodings such as RFC5198, and having
many tricks to reduce message overhead. The default UDP transport also forces
the protocol to reimplement some things that simulate point-to-point connection
channels within UDP.

Another option would have been modbus, which is significantly simpler than
CoAP, but the design of the data model shows its age a lot, and does not seem
too suited for embedded systems.

Instead of trying to add another complicated protocol, this specifies a tiny
protocol, that only has two primitives: Reading a block of memory and writing a
block of memory. This will allow the use of the register-table implementation
to be used as a memory backend, which allows specifying very flexible and
feature-rich remote memories.

This specifies the use of this protocol in two types of channels: TCP and
classical serial channels (think UART and RS232). In TCP channels, variable
integer length prefixing shall be used; in serial channels RFC1055 SLIP shall
be used.

All of the auxiliary technologies (register-table, varint-len-prefix, SLIP) are
already available in ufw.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>