ob: Add OctetChunks data type
In many situations when implementing communication systems, you have payload
sitting in one spot and meta data in another, and you want to transmit the meta
data as a header followed by the payload via a channel. In many systems you
need to be able to specify these chunks so sub-systems like framing protocols
can do their work.

This allows to build a table of memory chunks as OctetBuffer instances. There
is enough meta information included to use OctetChunks as a driver datum to
implement a Source instance on top of chunked data, which allows for easy use
with endpoint based systems like ufw's rfc1055 implementation or instance.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>