endpoints: Fix run-out behaviour of read_from_buffer()
Previously, the function would fail if there are not enough bytes in a buffer
to satisfy the request. But similar to APIs like POSIX's read(), it is also
possible to first drain the buffer completely and then fail when the function
is called with no additional data in the buffer.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>