Rename octet buffers to byte buffers
The backing data type here is an unsigned char, which is the perfect data type
for the byte of your architecture (whereas a byte is the smallest addressable
piece of memory on an architecture, rather than eight bits of data; while that
is the case on many architectures, there are some where it is not). I don't
quite remember when I chose "octet" before; "byte" is more accurate so we're
breaking API yet again.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>