rt: Remove size from serdes table
It is moved into its own table, since the data is used all over the place with-
out referencing the actual serdes at all. This will make it possible to use a
custom serdes for a given register table, since all the remaining uses of the
rds_serdes[] table now has access to the active register table as well.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>