ufw-sx: Add a small s-expression parser library
This uses malloc and is thus not part of ufw, the library. The parser can be
used to facilitate system testing. It was previously part of chip-remote's
system testing. Since some of the reusable parts of that are moved to ufwz, we
need this part either in ufwz or ufw. Since the library itself is very portable
with no dependency upon Zephyr, ufw itself is the more natural choice.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>