Add a toolchain file for afl++
It's not terribly difficult to use AFL++'s compilers, as shown by fuzz/run.

But the reason it is so easy for ufw is that it doesn't have any dependencies.
When makemehappy is the meta-build tool to resolve and build dependencies as
well, you need a toolchain file, because that is how makemehappy configures the
toolchains it is using. The afl++ toolchain file is not activated in by default
in makemehappy however.

It's not terribly difficult to configure a special purpose toolchain in mmh,
and the fuzz/afl++.yaml file is exactly one such example.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>