Add toolchain file for arm-zephyr-eabi
This allows using zephyr's ARM SDK compiler for arbitrary code. The toolchain
is basically a blessed GNU toolchain with a number of included batteries. It
works much like gnu-arm-none-eabi, and thus this automatically supports running
resulting binaries in QEMU.

To add this to mmh, you can do this in ~/.makemehappy/config.yaml (or whereever
in its configuration stack you prefer):

    toolchains:
      - name: arm-zephyr-eabi
        architecture: [ cortex-m3 ]

…adjust to taste.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>