gitint: Support prefix strings in version tags
By default the git-integration code matches tags that start in "v[0-9]*". In
some situations it may be useful to be able to use "foobar/v[0-9]*", to allow
for namespacing, with the "foobar/" prefix being inconsequential with regards
to the data-extraction.

This adds supports for that in build-information.sh and git.sh. On the CMake
side, this adds support for PREFIX in gitint_options() with a possible global
fallback of ${UFW_GIT_VERSION_PREFIX}.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>