regs: Add support to set and clear bits in registers
This allows the application to set and clear bits in registers without
probing the register type. Note that it is not supported to manipulate
bits for the registers of which type is either float or signed integer.
In such cases, the register is not modified and the user is notified
with the access error code, REG_ACCESS_INVALID.

Signed-off-by: Jiyeon Hong <jyeons.hong91@gmail.com>
Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>